WIP: CMakeLists: do not depend on vgrid
A rebuild from scratch fails when configuring iris, since it wants to
link with vgrid::vgrid-shared (via libgeoref), but the actual target made
available by vgrid is simply "vgrid::vgrid". Remove the dependency on
vgrid for now, while we untangle this CMake mess. This package is
actually not needed for the moment, and allows iris to be configured
correctly.