georef_cstintrp_reindexer: avoid ORCA halo columns indices
In the cstintrp weights for NEMO to GEM interpolation (which are on the
GEM grid), some GEM cells reference NEMO cells located in the first or
last NEMO columns. In NEMO, these columns are halo columms; the 1st
column corresponds to column 'jpiglo-1', and the last column (jpiglo)
corresponds to the second column.

Since in NEMO we define IRIS tiles that exclude these halo columns, we
end up interpolating uninitialized data since the 'UO' gridset
references these locations.

As such, modifiy the 'UO' gridset to reference the corresponding "real"
column when a NEMO cell has its "i" index in the first or last column.

We now need the dimension of the "other" grid also for 'UO' conversion,
so adjust the reference invocation.

Add an '--orca' flag to indicate that we want to account for ORCA halo
colums, and when given, set the i index to the "real" column when its
original value is in the first or last column.

Note that in the second ("Yang") loop, we add OtherDims[1] to the j
index, to account for the superposition of the Yin and Yang subgrid when
processing the GEM to NEMO interpolation weights (-g OU). We do this
addition unconditionally, which works because for NEMO to GEM (-g UO) we
do not invoke the reindexer with '-odim', and the default value for
'odim' is {0,0}, so we add zero in that case.

With the new behaviour introduced in this commit, we now pass '-odim'
for both invocations (OU and UO) but we want to shift the j index only
for '-g OU'. Add logic to that effect.