georef_cstintrp_reindexer: also output vector rotation factors ('#@@#')
The cstintrp weights files contain field 'ANG', which is the rotation
angle (on the destination grid) used to rotate vector field components
after interpolation. Components are rotated using:

    ur = cos(a)*u - sin(a)*v
    vr = sin(a)*u + cos(a)*v

where a is the angle from ANG multiplied by -pi/180.

Compute cos(a) and sin(a) and store them in a new field as
typvar='#@@#', so that these rotation factors are read into the 'R'
array when creating the Geoset object.

Since we treat the Yin and Yang subgrids in the same gridset, store the
sine and cosine values in a 1D array, i.e.
cos(a),sin(a),REF_INDEX_SEPARATOR,...,REF_INDEX_END.  This allows
storing both pairs of rotation factors for cells on the NEMO grid which
correspond to GEM cells in the Yin-Yang overlap region.

Remove the last section where we copy as-is the ANG field to the
destination file, since we now use '#@@#' instead.