GeoRef_Interp: apply rotation for vector components in GeoRef_InterpWeight
In a previous commit we added GeoSet->R, which holds the sine and cosine
of the grid rotation angle used to rotate vector field components to the
destination grid after interpolation (the ANG field in the cstintrp
weights). Adjust GeoRef_InterpWeight to actually use this field to
rotate each component after interpolation, but _before_ we average the
values from the Yin and Yang subgrids, in the Yin-Yang case.

Note that the '#@@#' field of the gridset (and thus the GeoSet->R array)
is structured as:

    cos(angle),sin(angle),REF_INDEX_SEPARATOR,...REF_INDEX_END.

The actual rotation follows what is done in CONCEPTS [1,2,3] and GEM
[4,5,6].

[1] https://gitlab.science.gc.ca/concepts/concepts/-/blob/d5ee63b933777c6b622eb9e8ed97c1c0286ae0a6/NEMOGCM/MY_SRC/sbcblk_rpnc.F90#L1436-1458
[2] https://gitlab.science.gc.ca/concepts/concepts/-/blob/d5ee63b933777c6b622eb9e8ed97c1c0286ae0a6/NEMOGCM/MY_SRC/sbcblk_rpnc.F90#L1999-2032
[3] https://gitlab.science.gc.ca/concepts/concepts/-/blob/d5ee63b933777c6b622eb9e8ed97c1c0286ae0a6/NEMOGCM/MY_SRC/sbcblk_rpnc.F90#L869-879
[4] https://gitlab.science.gc.ca/concepts/gem/-/blob/38796ad5c7914fd84016399af6872fc0e72f88c8/src/rpnphy/src/cpl/cplocn_xchng.F90#L214-236
[5] https://gitlab.science.gc.ca/concepts/gem/-/blob/b428f6bfe8705f7056ae2001d3e85fe83ed0846f/src/rpnphy/src/cpl/cplocn_xchng.F90#L267-284
[6] https://gitlab.science.gc.ca/concepts/gem/-/blob/b428f6bfe8705f7056ae2001d3e85fe83ed0846f/src/rpnphy/src/cpl/cplocn_distwgt.F90#L272-283