GeoRef_Interp: only rotate vector fields (!)
In b978b7f (GeoRef_Interp: apply rotation for vector components in
GeoRef_InterpWeight, 2025-08-27), we adjusted GeoRef_InterpWeight to use
the vector rotation factors in gset->R to correctly rotate vector field
components after interpolation. However, we do it unconditionnally, even
if interpolating a scalar field (i.e. even if zvin is NULL). This is
obviously wrong. Fix that by checking for 'zvin', as is done elsewhere
in the function, before computing the rotated components 'uvalr' and
'vvalr'. When zvin is null, simply set 'uvalr' to 'uval', which is
needed since we average Yin/Yang values afterwards using 'uvalr'. Note
that we still increment 'rp' unconditionnally since we also increment it
at the end of the loop to skip REF_INDEX_SEPARATOR.