GeoRef_InterpUV: clear both components prior to interpolation under IR_WEIGHTINDEX
GeoRef_InterpClear writes Opt->NoData to the entire destination field.
This is done in GeoRef_Interp but not in GeoRef_InterpUV, such that at
least for IR_WEIGHTINDEX the averaging in the Yin-Yang overlap region
does not work correctly since we potentially access uninitialized data.
Call GeoRef_InterpClear for both components prior to
GeoRef_InterpWeight.