georef_cstintrp_reindexer: skip point for which MASK is false
When comparing 1) the fields received by GEM from NEMO and interpolated
to the GEM grid in the GOSSIP solution and 2) the fields received by GEM
from NEMO already interpolated by IRIS (via libgeoref), we notice that
the land mask has more points in the IRIS solution [1]. The reindexer
does not take into account the MASK field, and creates the gridset by
including the points where NAVG is nonzero.

Since NAVG is a masked field, librmn can (and does) write non-zero
values at masked points near the border of the mask, for compression
efficiency. This means that we include masked points in the gridset
because NAVG is non-zero at those points.

We already read the MASK field since a few commits ago, so add a
condition in the Yin and Yang loops to check if the point is masked, and
exclude it from the gridset if so (by setting in=FALSE). This makes the
UO gridset (NEMO to GEM) a bit smaller i.e. less points are considered
ocean (and thus more points are considered land). With this change to
the UO gridset we get the same mask in IRIS as in the GOSSIP solution.

[1] https://gitlab.science.gc.ca/phc001/gem-iris-nemo/-/issues/4#note_1228079