georef_cstintrp_reindexer: fix ORCA halo columns adjustment for Yang grid
Since e52ec81 (georef_cstintrp_reindexer: avoid ORCA halo columns
indices, 2025-08-22), when creating the gridset for UO interpolation
(NEMO to GEM), we adjust the i indices of NEMO points located in the
first and last column, since these columns are halo columns and are not
sent to IRIS. We thus use the corresponding point in the penultimate
column, for points in the first column, and the corresponding point in
the second column, for points in the last column.

At least, this is what we do for the Yin subgrid. The corresponding code
for the Yang subgrid uses the corresponding point in the penultimate
column _if 'iy==1'_, which is wrong because iy is set to
'iy_data[1][n][idx]-1', i.e. the i index read in the cstintrp weight
file, minus 1. We substract one since the indices in the cstintrp
weights files are Fortran indices, thus 1-based, but the reindexer
creates gridsets for libgeoref, which is in C and thus 0-based. 'iy' is
thus a 0-based index, and as such its value in the first column is 0,
not 1. This means that we end up modifying the i indices for points in
the second column instead of the first.

Fix the condition such that we correctly adjust the i index for points
in the first column (iy==0).