Prioritize installed CMakeRPN
We will stop updating the cmake_rpn submodule although we will leave it there to
provide a convenient way to get it.  Because we don't update it anymore, getting
the latest version can be done with `git submodule update --init --remote`.

Instructions have been added to the readme recommending this approach or to
install it and use `EC_CMAKE_MODULE_PATH` to provide it to other EC projects.